Introduce graphql, data, and domain layers
