+++ src/main/kotlin/com/lange/biblioteque/data/RepositoryFactory.kt
@@ -1,5 +1,6 @@
-package com.lange.biblioteque.plugins
+package com.lange.biblioteque.data
 
+import com.lange.biblioteque.data.openlibrary.OpenLibraryRepository
 import io.ktor.client.*
 import io.ktor.client.engine.cio.*
 import io.ktor.client.plugins.contentnegotiation.*
@@ -7,7 +8,10 @@ import io.ktor.client.plugins.logging.*
 import io.ktor.serialization.kotlinx.json.*
 import kotlinx.serialization.json.Json
 
-object Client {
+object RepositoryFactory {
+    val openLibraryRepository: OpenLibraryRepository
+        get() = OpenLibraryRepository(client = clientInstance)
+
     val clientInstance: HttpClient
         get() = HttpClient(CIO) {
             install(Logging)
