+++ src/main/kotlin/com/lange/biblioteque/graphql/BookSchema.kt
@@ -1,9 +1,13 @@
-package com.lange.biblioteque.api.v1.isbn
+package com.lange.biblioteque.graphql
 
 import com.apurebase.kgraphql.schema.dsl.SchemaBuilder
-import com.lange.biblioteque.client
+import com.lange.biblioteque.data.RepositoryFactory
+import com.lange.biblioteque.data.openlibrary.OpenLibraryRepository
+import com.lange.biblioteque.domain.book.*
 
-fun SchemaBuilder.bookSchema() {
+fun SchemaBuilder.bookSchema(
+    openLibraryRepository: OpenLibraryRepository = RepositoryFactory.openLibraryRepository
+) {
     type<BookDataResponse> {
         description = "Information about a book"
     }
@@ -35,7 +39,7 @@ fun SchemaBuilder.bookSchema() {
     query("book") {
         description = "Returns information about a book"
         resolver { isbn: String ->
-            client.getBookDataViaIsbn(isbn = isbn)
+            openLibraryRepository.getBookDataViaIsbn(isbn = isbn)
         }
     }
 }
