+++ src/main/kotlin/com/lange/biblioteque/plugins/GraphQl.kt
@@ -1,7 +1,7 @@
 package com.lange.biblioteque.plugins
 
 import com.apurebase.kgraphql.GraphQL
-import com.lange.biblioteque.api.v1.isbn.bookSchema
+import com.lange.biblioteque.graphql.bookSchema
 import io.ktor.server.application.*
 
 fun Application.configureGraphQL() {
