--- src/main/kotlin/com/lange/biblioteque/plugins/GraphQlSchema.kt
@@ -1,2 +1,0 @@
-package com.lange.biblioteque.plugins
-
